SEQNG-773 Some refactorings
